askfaithbot-rag-lambda/requirements.txt
boto3
botocore
python-dotenv
langchain
langchain-aws
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
